[extract.sh] improve extract script

+++ extract.sh
@@ -210,10 +210,14 @@ bash tools/misc/compose.sh ${PWD}/output/examples/linkkit_example_auto.c
 echo "Please pick up extracted source files in [${PWD}/${OUTPUT_DIR}]"
 echo ""
 
-cp tools/misc/makefile.output output/Makefile
-ENV_TEST=$(cat .config 2>/dev/null| sed -n '/VENDOR/{s/[[:space:]]//gp}'| awk -F ':' '{print $2}')
-if [ "${ENV_TEST}" = "ubuntu" ];then
-    rm -f ${WRAPPERS_DIR}/wrapper.c
-    cp -rf wrappers/os/ubuntu ${WRAPPERS_DIR}/
-    rm -f ${WRAPPERS_DIR}/ubuntu/HAL_UART_linux.c
+if [ "${1}" = "test" ];then
+    cp tools/misc/makefile.output output/Makefile
+    ENV_TEST=$(cat .config 2>/dev/null| sed -n '/VENDOR/{s/[[:space:]]//gp}'| awk -F ':' '{print $2}')
+    if [ "${ENV_TEST}" = "ubuntu" ];then
+        rm -f ${WRAPPERS_DIR}/wrapper.c
+        cp -rf wrappers/os/ubuntu ${WRAPPERS_DIR}/
+        cp -rf wrappers/tls ${WRAPPERS_DIR}/
+        cp -rfl external_libs ${WRAPPERS_DIR}/
+        rm -f ${WRAPPERS_DIR}/ubuntu/HAL_UART_linux.c
+    fi
 fi
